change mail templates, draw bar chart

+++ src/main/java/com/thd/pos_system_java_final/controllers/ProductController.java
@@ -1,6 +1,7 @@
 package com.thd.pos_system_java_final.controllers;
 
 import com.thd.pos_system_java_final.models.Cart.Item;
+import com.thd.pos_system_java_final.models.Order.OrderDetailRepository;
 import com.thd.pos_system_java_final.models.Product.Product;
 import com.thd.pos_system_java_final.models.Product.ProductRepository;
 import com.thd.pos_system_java_final.services.CartService;
@@ -38,6 +39,8 @@ public class ProductController {
 
     @Autowired
     private CartService cartService;
+    @Autowired
+    private OrderDetailRepository orderDetailRepository;
 
     @GetMapping("")
     public String listProduct(Model model) {
@@ -111,7 +114,7 @@ public class ProductController {
     @PostMapping("/delete/{id}")
     public String delete(@PathVariable int id) {
         List<Item> items = (List<Item>) session.getAttribute("cart");
-        if (!cartService.isInOrder(id)) {
+        if (!cartService.isInOrder(id) && !orderDetailRepository.existsByProductId(id)) {
             productRepository.deleteById(id);
         }
